reactor-webrtc: don't narrow the 64-bit packet counters
Found by the automated review on #80: the candidate pair's `packets_sent` and
`packets_received` are `std::optional<uint64_t>` in this libwebrtc, and I was
casting them into the entry's `uint32_t` fields to reuse them.

Two more of the same shape were already there, in the lines this PR is editing:
`RTCSentRtpStreamStats::packets_sent` and
`RTCOutboundRtpStreamStats::retransmitted_packets_sent` are both `uint64_t` and
have been narrowed implicitly since the fields existed. A connection carrying a
thousand packets a second passes 2^32 in about seven weeks, and then a
cumulative counter appears to go backwards — the one thing a cumulative counter
must not do. Fixing one and leaving two in the same function would have been a
worse outcome than the widening.

So `packets_sent`, `retransmitted_packets_sent` and the pair's own
`packets_received` are 64-bit through the whole path.
`InboundRtpStats::packets_received` stays `u32`, because
`RTCReceivedRtpStreamStats` genuinely reports 32 bits there — the pair's count
gets its own field rather than sharing that one.

`OutboundRtpStats::packets_sent` and `retransmitted_packets_sent` widening from
`u32` is this release's only breaking change, and the CHANGELOG now says so
instead of claiming the release is purely additive. It is a widening, so a
caller binding to `u64` is unaffected; one binding to `u32` is not. The known
consumers: reactor-runtime reads these into Python ints and is unaffected;
reactor-client-sdks holds `retransmitted_packets_sent` as `u32` and will need
the same widening when it takes 0.15.

The struct's size assertions caught my own arithmetic twice while doing this —
first an off-by-one in the 4-byte block, then a redundant pad — which is the
argument for having them.

+++ crates/reactor-webrtc-py/src/lib.rs
@@ -847,7 +847,9 @@ pub struct OutboundRtpStats {
     pub ssrc: u32,
     /// Audio or video.
     pub kind: StreamKind,
-    pub packets_sent: u32,
+    /// 64-bit because libwebrtc reports it that way; a 32-bit counter wrapped
+    /// silently on a long-lived connection.
+    pub packets_sent: u64,
     pub bytes_sent: u64,
     /// Target encoder bitrate in bps.
     pub target_bitrate_bps: f64,
@@ -862,7 +864,7 @@ pub struct OutboundRtpStats {
     pub fraction_lost: f64,
     /// Packets the receiver reports as lost. Signed, per RFC 3550.
     pub packets_lost: i32,
-    pub retransmitted_packets_sent: u32,
+    pub retransmitted_packets_sent: u64,
     /// Encoded frames per second; `0.0` if not measured. Video only.
     pub frames_per_second: f64,
     pub frames_sent: u32,
@@ -920,8 +922,8 @@ pub struct IceCandidatePairStats {
     /// than the per-stream RTP counters.
     pub bytes_sent: u64,
     pub bytes_received: u64,
-    pub packets_sent: u32,
-    pub packets_received: u32,
+    pub packets_sent: u64,
+    pub packets_received: u64,
     /// Type of this pair's *local* candidate. `RELAY` says the session is going
     /// through TURN.
     pub local_candidate_type: IceCandidateType,
